+++ packages/donation-form/src/form-elements/contact-form.ts
@@ -7,8 +7,8 @@ import {
   TemplateResult,
   query
 } from 'lit-element';
-import { BillingInfo } from './models/common/billing-info';
-import { CustomerInfo } from './models/common/customer-info';
+import { BillingInfo } from '../models/common/billing-info';
+import { CustomerInfo } from '../models/common/customer-info';
 
 @customElement('contact-form')
 export class ContactForm extends LitElement {
